Create ScoringTheRace
The "ScoringTheRace" program takes race outcome inputs and analyzes them to provide perspectives into team performance. It calculates team scores based on the order of runners' finish, determining the number of participating teams and the size of the largest team. Furthermore, it confirms if all teams have an equal number of runners. The program then outputs the number of participating teams, the size of each team, and their average score. It identifies the winning team(s) by comparing their total scores, declaring the team(s) with the lowest score(s) as the winner(s). This iterative program enables users to analyze multiple race outcomes effortlessly.