Create TowerOfHanoi
"TowerOfHanoiSolver" is a program created to effectively solve the Tower of Hanoi puzzle. When given the number of rings, it figures out the best order of moves needed to transfer all rings from one pole to another, following the rules of the puzzle. It's a tool that offers a great solution to this well-known problem. It helps users understand and navigate it easily.